requirements.txt
streamlit
streamlit-lottie
plotly
requests
torch==2.4.1
torchvision==0.19.1
transformers==4.45.1
timm==0.9.12
opencv-python-headless
Pillow
matplotlib
pycocotools
numpy
streamlit-option-menu
supervision>=0.22.0
addict
yapf
opencv-python